+++ changelog.md
@@ -6,3 +6,5 @@
 
 - Added wn8 calculation
 - Added update checking at app runtime
+- Added caling tomato.gg api to get player sessions
+- Added history of nicknames searched